+++ com.unity.renderstreaming/Runtime/Scripts/RenderStreaming.cs
@@ -113,7 +113,6 @@ namespace Unity.RenderStreaming
                 conf.GetValueOrDefault(new RTCConfiguration { iceServers = iceServers });
             if (hardwareEncoder != null)
                 hardwareEncoderSupport = hardwareEncoder.Value;
-            var encoderType = hardwareEncoderSupport ? EncoderType.Hardware : EncoderType.Software;
 
             if (signaling != null)
             {
@@ -128,7 +127,6 @@ namespace Unity.RenderStreaming
             RenderStreamingDependencies dependencies = new RenderStreamingDependencies
             {
                 config = _conf,
-                encoderType = encoderType,
                 signaling = _signaling,
                 startCoroutine = StartCoroutine,
                 resentOfferInterval = interval,