+++ com.unity.renderstreaming/Runtime/Scripts/RenderStreamingInternal.cs
@@ -22,11 +22,6 @@ namespace Unity.RenderStreaming
         /// <summary>
         ///
         /// </summary>
-        public EncoderType encoderType;
-
-        /// <summary>
-        ///
-        /// </summary>
         public RTCConfiguration config;
 
         /// <summary>
@@ -115,7 +110,7 @@ namespace Unity.RenderStreaming
 
             if (s_list.Count == 0)
             {
-                WebRTC.WebRTC.Initialize(dependencies.encoderType);
+                WebRTC.WebRTC.Initialize();
             }
 
             _config = dependencies.config;
@@ -226,11 +221,6 @@ namespace Unity.RenderStreaming
             var peer = _mapConnectionIdAndPeer[connectionId];
             RTCRtpSender sender = peer.peer.AddTrack(track);
             var transceiver = peer.peer.GetTransceivers().First(t => t.Sender == sender);
-
-            // note:: This line is needed to stream video to other peers with hardware codec.
-            // The exchanging SDP is failed if remove the line because the hardware decoder currently is not supported.
-            // Please remove the line after supporting the hardware decoder.
-            transceiver.Direction = RTCRtpTransceiverDirection.SendOnly;
             return transceiver;
         }