+++ com.unity.renderstreaming/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to com.unity.renderstreaming package will be documented in t
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - yyyy-mm-dd
+
+### Changed
+
+- Upgraded the version of WebRTC package `2.4.0-exp.7`.
+
 ## [3.1.0-exp.3] - 2022-03-01
 
 ### Changed